refactor(tests): consolidate duplicated test helpers into shared modules (#1028)
Moves repeated test fixture functions out of individual test files and into
the per-crate common/ modules, eliminating SonarCloud duplication hits.

lib-consensus:
- consensus_fixtures: add non_zero_key(), zero_key(), nonce_signature(),
  network_summary() — covers all remaining local helpers
- security_fixes_tests: redirect create_test_key/signature to fixtures
- heartbeat_tests: redirect create_test_network_summary to fixtures

lib-blockchain:
- block_builders: add test_input(), test_output(), test_transaction_at()
- sov_week7: add mod common; redirect signature/pubkey/input/output helpers
- sov_week10: add mod common; redirect pubkey/transaction helpers via super::common

lib-storage:
- Create tests/common/storage_fixtures with test_backend() and test_config()
- storage_load_tests: redirect create_test_backend to fixtures
- dht_nodeid_stability: redirect create_test_config to fixtures

Closes #1028
