+++ lib-storage/tests/common/mod.rs
@@ -1,0 +1,3 @@
+//! Shared test helpers for lib-storage integration tests.
+
+pub mod storage_fixtures;
